large-file: advertise CAP_LARGE_FILES + 64-bit size plumbing
Wires up the Capabilities-Large-File extension per fogWraith/Hotline
Docs/Protocol/Capabilities-Large-File.md. Control plane and HTXF
handshake are complete; data-plane FFO fork-header 64-bit split
encoding is a follow-up.

Capability negotiation
  * Advertise CAP_LARGE_FILES (bit 0) alongside CAP_TEXT_ENCODING
    in the LOGIN DATA_CAPABILITIES chunk.
  * Echo recognition: rcv_task_login now logs "server confirmed
    large-file (64-bit) mode" when the server echoes the bit back.

New wire constants (hotline.h)
  * HTLC/HTLS_DATA_FILESIZE64        0x01f1   (8-byte BE)
  * HTLC/HTLS_DATA_OFFSET64          0x01f2
  * HTLC/HTLS_DATA_XFERSIZE64        0x01f3
  * HTLC/HTLS_DATA_FOLDER_ITEM_COUNT64 0x01f4
  * HTXF_FLAG_LARGE_FILE             0x00000001
  * HTXF_FLAG_SIZE64                 0x00000002

Internal types
  * struct htxf_conn size/position fields (data_size, data_pos,
    rsrc_size, rsrc_pos, total_size, total_pos, srv_data_size)
    bumped guint32 → guint64.
  * human_readable / human_size now take guint64. gtkhx_session
    file-info signal payload bumped to G_TYPE_UINT64. Display
    formatters updated.

Receive side (rcv.c)
  * rcv_task_file_get: parses HTLS_DATA_XFERSIZE64, prefers it
    over the legacy 32-bit DATA_HTXF_SIZE when present.
  * rcv_task_folder_get: same, plus accepts XFERSIZE64 for the
    aggregate folder tree byte count.
  * rcv_task_file_getinfo: parses HTLS_DATA_FILESIZE64 and emits
    it on the file-info signal.

Send side (xfers.c)
  * xfer_go for FILE_PUT: when CAP_LARGE_FILES is active, sends
    both the clamped 32-bit DATA_HTXF_SIZE AND the 8-byte
    DATA_XFERSIZE64. Legacy peers ignore the 64-bit chunk;
    large-file peers prefer it.

HTXF handshake (network.c)
  * htxf_connect detects CAP_LARGE_FILES via htxf->htlc->caps and:
      - sets HTXF_FLAG_LARGE_FILE in the flags low byte
      - if total_size > 0xFFFFFFFF, also sets HTXF_FLAG_SIZE64,
        zeros the legacy 32-bit length, and appends an 8-byte BE
        length to make the 24-byte handshake variant
  * The historical "type in high u16 of unknown" encoding lives
    in the high half; new flags occupy the low half. The two
    interpretations co-exist on the wire so a Mac-native server
    reading type still gets the right value, and a large-file
    server reading flags also gets the right value.

Tests
  * Existing test_capabilities covered CAP_LARGE_FILES via the
    multi-bit cases; no changes.
  * New tests/proto/test_large_file.c pins the new field IDs,
    the HTXF flag constants, send/recv shapes for DATA_XFERSIZE64,
    and the three handshake variants (legacy 16-byte, large-file
    < 4 GiB 16-byte with LARGE_FILE flag, large-file > 4 GiB
    24-byte with both flags + appended 8-byte length).

Deferred (follow-up)
  * FFO fork-header 64-bit split encoding (Capabilities-Large-File
    spec, section "Flattened File Object Fork Headers"): the
    16-byte fork header repurposes the Compression field (offset
    4-7) as the high 32 bits of the fork length when the
    HTXF_FLAG_LARGE_FILE flag is set. Touches file_send_one /
    file_recv_one in the xfer worker thread. Until that lands,
    downloads of files > 4 GiB are not supported (FFO encode
    truncates), but the handshake-level support, capability
    negotiation, and uploads (raw data, no FFO wrapper per spec)
    work.

TODO: remove "Support Download Info (211)" — we already do (opcode
0xD3 = HTLS_HDR_QUEUE in our code, with the same REFNUM + waiting
count fields; Mobius just names it differently).

All 41 unit + proto tests pass.

+++ src/files.h
@@ -46,7 +46,7 @@ struct hl_filelist_hdr;
  * is the receiving end that builds the dialog. */
 extern void output_file_info (char *path, char *name, char *creator, char *type,
                               char *comments, char *modified, char *created,
-                              guint32 size);
+                              guint64 size);
 
 /* Emit the file-list GtkhxSession signal so the file_list-signal
  * handler in gtkhx.c (and through it the remote-provider in the
